+++ src/main.cpp
@@ -9,7 +9,7 @@ int main() {
     for (auto [factory, total_days, day_id] = std::tuple(Factory{}, 25, 1);
          day_id <= total_days; ++day_id)
     {
-        std::unique_ptr<Day>day {factory.createDay(day_id)};
+        std::unique_ptr<Day>day {factory.createDay(static_cast<int8_t>(day_id))};
         day->print_header();
         day->calculate_all();
         day->print_results();
